refactor(anthropic-vertex): move sdk runtime to plugin
